requirements.txt
# Core dependencies
apache-beam[gcp]==2.50.0
google-cloud-pubsub==2.13.11
google-cloud-bigquery==3.5.0
google-cloud-storage==2.10.0
pytest==7.4.0
pyyaml==6.0.1
google-cloud-dataflow-client==0.32.0  # Official client library

# Development dependencies
flake8==6.1.0
black==23.12.1
isort==5.13.2
mypy==1.8.0

# Additional dependencies
requests==2.31.0
pandas==2.1.4
numpy==1.24.3  # Apache-beam compatible
